Avoid allocating String for datum time
A huge amount of the parse time is spent allocating Strings. The string
associated with the time of each datum is never used as a string but
just parsed to a u64. This commit avoids allocating those as strings in
the first place.

I used the following program to generate a 1.6 gigabyte sample input
that resembles what you show in your presentation.
https://www.youtube.com/watch?v=aWbGPMxs0AM

    for i in 0..30000000 {
        let time = 20000000000u64 + i;
        let entity = 50000 + i % 1000;
        let state = i % 3;
        println!(r#"{{ "time": "{}", "entity": "{}", "state": {} }}"#, time, entity, state);
    }

On this input, this commit drops parse time from 6.8 seconds to 5.7
seconds. I isolated parse time by inserting an early return just below
the line that contains "let nstates: u32".